Record API, Mission API 구현
- Record 글CRUD 구현
- MemberMissionRepository 만들기 위해 MemberMissionId DTO 생성
- Mission 생성/참여/탈퇴 시 MemberMission 테이블 update

- Mission 탈퇴시 시작일 이전인지, 미션 호스트인지 체크하는 부분
  추가해야함

+++ back/seesaw/src/main/java/com/doacha/seesaw/model/entity/Record.java
@@ -18,7 +18,7 @@ public class Record {
     @Id
     @Column(name = "record_id")
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int recordId;
+    private Long recordId;
 
     @Column(name = "record_content")
     private String recordContent;
@@ -41,6 +41,4 @@ public class Record {
     @JoinColumn(name="mission_id",referencedColumnName = "mission_id",nullable = false),
     @JoinColumn(name="member_email", referencedColumnName = "member_email",nullable = false)})
     private MemberMission memberMission;
-
-
 }
